+++ index.d.ts
@@ -22,6 +22,10 @@ declare namespace AlfrescoApi {
         upload: any;
         webScript: any;
 
+        ecmClient : EcmClient;
+        bpmClient : BpmClient;
+        ecmPrivateClient : EcmPrivateClient;
+
         changeEcmHost(ecmHost: string): void;
         changeBpmHost(bpmHost: string): void;
         changeCsrfConfig(disableCsrf: boolean): void;
@@ -145,6 +149,24 @@ declare namespace AlfrescoApi {
         content: ContentInfo;
         path: PathInfoEntity;
         properties: NodeProperties;
+    }
+
+    export interface EcmClient {
+        new(config: AlfrescoApiConfig): EcmClient;
+        changeHost(host: String): void;
+        setAuthentications(authentications: any): void;
+    }
+
+    export interface BpmClient {
+        new(config: AlfrescoApiConfig): BpmClient;
+        changeHost(host: String): void;
+        setAuthentications(authentications: any): void;
+    }
+
+    export interface EcmPrivateClient {
+        new(config: AlfrescoApiConfig): EcmPrivateClient;
+        changeHost(host: String): void;
+        setAuthentications(authentications: any): void;
     }
 
     export interface MinimalNodeEntryEntity extends MinimalNode {
@@ -354,6 +376,7 @@ declare namespace AlfrescoApi {
         sitesApi: any;
         queriesApi: any;
         tagsApi: any;
+        webscriptApi: any;
 
         /*Models*/
         Activity: any;
@@ -471,6 +494,7 @@ declare namespace AlfrescoApi {
         hostEcm: string;
         hostBpm: string;
         contextRoot: string;
+        contextRootBpm: string;
         provider: string;
         ticketEcm: string;
         ticketBpm: string;