+++ src/ecmAuth.js
@@ -23,8 +23,7 @@ class EcmAuth extends AlfrescoApiClient {
         Emitter.call(this);
     }
 
-    changeHost(host) {
-        this.config.hostEcm = host;
+    changeHost() {
         this.basePath = this.config.hostEcm + '/' + this.config.contextRoot + '/api/-default-/public/authentication/versions/1'; //Auth Call
         this.ticket = undefined;
     }