+++ CHANGELOG.md
@@ -9,6 +9,7 @@ _This project provides a JavaScript client API into the v1 Alfresco REST API_
 <a name="1.0.1"></a>
 # [1.0.1](https://github.com/Alfresco/alfresco-js-api/releases/tag/1.0.0) (xx-xx-2017)
 ## fix
+- [Ticket not cleaned when host is changed #143](https://github.com/Alfresco/alfresco-js-api/pull/143)
 - [Wrong PathInfoEntity declaration #139](https://github.com/Alfresco/alfresco-js-api/pull/139)
 
 <a name="1.0.0"></a>