+++ CHANGELOG.md
@@ -5,6 +5,10 @@
 Alfresco JS API
 
 _This project provides a JavaScript client API into the v1 Alfresco REST API_
+<a name="1.0.0"></a>
+# [1.0.0](https://github.com/Alfresco/alfresco-js-api/releases/tag/1.0.0) (20-01-2017)
+## fix
+- [Wrong API name - getRawContent3 #156](https://github.com/Alfresco/alfresco-js-api/issues/156)
 
 <a name="1.0.1"></a>
 # [1.0.1](https://github.com/Alfresco/alfresco-js-api/releases/tag/1.0.0) (xx-xx-2017)